+++ crates/noseyparker-cli/tests/scan/mod.rs
@@ -7,4 +7,9 @@ mod with_ignore;
 // TODO: add test for scanning with `--github-user`
 // TODO: add test for scanning with `--github-org`
 // TODO: add test for caching behavior of rescanning `--git-url`
+// TODO: add test for scanning multiple times with changing `--git-clone-mode` option
+// TODO: add test for scanning with `--git-clone-mode bare` and `--git-clone-mode mirror`
+// TODO: add test for scanning with `--github-api-url`
 // TODO: add tests for SARIF output format
+// TODO: add test for scanning with non-default `--snippet-length` parameter
+// TODO: add test for scanning multiple times with changing values of `--snippet-length`